requirements.txt
asgiref==3.7.2
Django==4.2.3
django-bootstrap5==23.3
django-filter==23.2
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
